requirements.txt
discord.py
pillow
asyncio
requests


^^^^^^^^^^^

prolly some other ones i cba